oncor-dashboard/requirements.txt
dash>=2.18.0
dash-bootstrap-components>=1.6.0
plotly>=5.24.1
pandas>=2.3.0
requests>=2.32.0
pdfplumber>=0.11.0
python-dateutil>=2.9.0
